Ainda em entidades

+++ Desenhador.h
@@ -12,8 +12,8 @@ class Sprite{
 public:
     ///Cria um sprite a partir de um arquivo texto, posições no espaço (x e y) e ordem em que vai ser desenhado (profundidade z)
     Sprite();
-    Sprite(std::string &arquivo, int x, int y, int z);
-    Sprite(std::vector <std::vector <char>> &sprite, int x, int y, int z);
+    Sprite(std::string &arquivo, float x, float y, int z);
+    Sprite(std::vector <std::vector <char>> &sprite, float x, float y, int z);
 
     void setSprite(std::vector <std::vector <char>> &sprite);
     void setSprite(std::string &arquivo);
@@ -26,10 +26,10 @@ public:
     size_t H();
 
     ///Posição horizontal
-    int x;
+    float x;
 
     ///Posição vertical
-    int y;
+    float y;
 
     ///Profundidade ou ordem em que será desenhado
     int z;
